Define klein's behavior in terms of request.write being called after request.finish.
This is caused by either the client terminating the connection or a bug in the server.

The default behavior is to allow request.write to be called when a handler completes
successfully, and let twisted.web raise it's RuntimeError.

However, the processingFailed code bath has been changed to guard against
error handling being called when a request has already completed.

This also allows this pattern:

@route("/")
def handler(request):
    d = someOperation()
    request.notifyFinish().addCallback(lambda _: d.cancel())

    return d

Which will effectively cancel someOperation when the client the disconnects,
while the handler is waiting for the result of someOperation.