Improve testing instructions and flexibility.
The directions for contributing now go into detail about how to run the
tests, to emphasize that testing is important. As well, tox.ini has been
amplified by using posargs: its default behavior is the same as before,
but now it can be given arguments to run only a subset of Klein's tests,
allowing for faster testing when in development.