consistently make 'style' the _outside_ decorator
for most of these, it should still work either way, but ultimately we
will have validation-failure logic which is at the "bottom" of the route
stack because there is actually no route that underlies it; it's a part
of a previously-declared route.  given that one form is going to have to
be on the 'inside' here, we should make it consistent.

also, less indentation on the templates looks nicer.