Make widget tests fail with free and bound functions.
Widgets are rendered by invoking a `widgeted` function patched onto
their rendering function.  When the render function is a bound method,
it can't forward `self` to the invocation of the render function
inside `widgeted`.

Also, because `widgeted`'s signature assumes an instance, render
functions that do not have a first positional argument cannot be used
with widgets.

The tests have been modified to fail on these two cases.