Plated routes can serialize slots with Deferreds to JSON.
Before this patch, these could be rendered as HTML but would fail with
a TypeError if JSON was requested.

klein._plating.resolve_deferred_objects traverses a JSON
serializable-object, waiting on any Deferreds it may contain, or it
itself if it's a Deferred, and returns a new, wholly resolved result.

Plating.routed calls this instead of json_serialize to obtain slot
data that can be serialized to JSON.