Move zope interfaces, types, and exceptions from _headers.py/_request.py/_response.py to _imessage.py. That will enable imports o fthe interfaces without the implementations.
Move _interfaces.py to _iapp.py, since the one interface there (IKleinRequest) is associated with app.py.

Make a new _interfaces.py which imports the zope interfaces from all of the above and either exports them directly or, when type checking, exports types that make mypy happier.
This is a work-around until there is a Zope Interface plugin for mypy.